# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-SERVICES marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  role: string;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  phone?: string | null;
  isVerified?: boolean;
  isBanned?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  name?: string;
  bio?: string;
  location?: string;
  phone?: string;
  avatarUrl?: string;
}

export interface FreelancerProfile {
  id: number;
  userId: number;
  level: string;
  skills?: string;
  totalEarnings: number;
  completedOrders: number;
  rating: number;
  reviewCount: number;
  isAvailable?: boolean;
  createdAt: string;
  user?: User;
}

export interface FreelancerProfileInput {
  skills: string;
  bio?: string;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  /** @nullable */
  description?: string | null;
  serviceCount: number;
}

export interface CategoryInput {
  name: string;
  icon: string;
  description?: string;
}

export interface CategoryUpdate {
  name?: string;
  icon?: string;
  description?: string;
}

export interface Service {
  id: number;
  title: string;
  description: string;
  price: number;
  deliveryDays: number;
  categoryId: number;
  freelancerId: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  tags?: string | null;
  status: string;
  orderCount: number;
  rating: number;
  reviewCount: number;
  isFeatured?: boolean;
  createdAt: string;
  category?: Category;
  freelancer?: FreelancerProfile;
}

export interface ServiceInput {
  title: string;
  description: string;
  price: number;
  deliveryDays: number;
  categoryId: number;
  imageUrl?: string;
  tags?: string;
}

export interface ServiceUpdate {
  title?: string;
  description?: string;
  price?: number;
  deliveryDays?: number;
  categoryId?: number;
  imageUrl?: string;
  tags?: string;
  status?: string;
}

export interface ServiceListResponse {
  services: Service[];
  total: number;
  page: number;
  limit: number;
}

export interface FavoriteStatus {
  isFavorited: boolean;
}

export interface Order {
  id: number;
  serviceId: number;
  clientId: number;
  freelancerId: number;
  status: string;
  amount: number;
  commission: number;
  freelancerAmount: number;
  description: string;
  /** @nullable */
  attachmentUrl?: string | null;
  /** @nullable */
  deliveryUrl?: string | null;
  deliveryDays: number;
  deadline: string;
  /** @nullable */
  revisionNote?: string | null;
  createdAt: string;
  updatedAt?: string;
  service?: Service;
  client?: User;
  freelancer?: FreelancerProfile;
}

export interface OrderInput {
  serviceId: number;
  description: string;
  attachmentUrl?: string;
}

export interface DeliveryInput {
  deliveryUrl: string;
  note?: string;
}

export interface RevisionInput {
  note: string;
}

export interface OrderStats {
  totalOrders: number;
  pendingOrders: number;
  inProgressOrders: number;
  completedOrders: number;
  cancelledOrders: number;
  totalRevenue: number;
}

export interface CheckoutInput {
  orderId: number;
}

export interface CheckoutResponse {
  paymentUrl: string;
  reference: string;
}

export interface Payment {
  id: number;
  orderId: number;
  amount: number;
  status: string;
  reference: string;
  provider?: string;
  createdAt: string;
}

export interface WebhookResponse {
  received: boolean;
}

export interface Review {
  id: number;
  orderId: number;
  serviceId: number;
  clientId: number;
  freelancerId: number;
  rating: number;
  comment: string;
  createdAt: string;
  client?: User;
}

export interface ReviewInput {
  orderId: number;
  rating: number;
  comment: string;
}

export interface Message {
  id: number;
  orderId: number;
  senderId: number;
  content: string;
  /** @nullable */
  attachmentUrl?: string | null;
  createdAt: string;
  sender?: User;
}

export interface MessageInput {
  content: string;
  attachmentUrl?: string;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  title: string;
  body: string;
  /** @nullable */
  link?: string | null;
  isRead: boolean;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  totalFreelancers: number;
  totalServices: number;
  totalOrders: number;
  totalRevenue: number;
  platformCommission: number;
  pendingOrders: number;
  completedOrders: number;
}

export interface SuccessResponse {
  success: boolean;
}

export type ListServicesParams = {
q?: string;
categoryId?: number;
minPrice?: number;
maxPrice?: number;
sort?: string;
page?: number;
limit?: number;
};

export type ListOrdersParams = {
status?: string;
role?: string;
};

export type AdminListUsersParams = {
page?: number;
limit?: number;
q?: string;
};