# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: E-SERVICES marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    patch:
      operationId: updateCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: No content

  # SERVICES
...
  /messages/{orderId}:
    get:
      operationId: listMessages
      tags: [messages]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      parameters:
        - name: orderId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"

  # NOTIFICATIONS
...
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      parameters:
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
        - name: q
          in: query
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/ban:
    post:
      operationId: adminBanUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/services/{id}/approve:
    post:
      operationId: adminApproveService
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name: { type: string, minLength: 2 }
        email: { type: string }
        password: { type: string, minLength: 6 }
        role: { type: string }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token: { type: string }
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        avatarUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        location: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        isVerified: { type: boolean }
        isBanned: { type: boolean }
        createdAt: { type: string }
...
    FreelancerProfile:
      type: object
      required: [id, userId, level, totalEarnings, completedOrders, rating, reviewCount, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        level: { type: string }
        skills: { type: string }
        totalEarnings: { type: number }
        completedOrders: { type: integer }
        rating: { type: number }
        reviewCount: { type: integer }
        isAvailable: { type: boolean }
        createdAt: { type: string }
        user:
          $ref: "#/components/schemas/User"
...
    Category:
      type: object
      required: [id, name, slug, icon, serviceCount]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        icon: { type: string }
        description: { type: ["string", "null"] }
        serviceCount: { type: integer }
    CategoryInput:
      type: object
      required: [name, icon]
      properties:
        name: { type: string }
        icon: { type: string }
        description: { type: string }
    CategoryUpdate:
      type: object
      properties:
        name: { type: string }
        icon: { type: string }
        description: { type: string }

    # SERVICE
    Service:
      type: object
      required: [id, title, description, price, deliveryDays, categoryId, freelancerId, status, orderCount, rating, reviewCount, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: string }
        price: { type: number }
        deliveryDays: { type: integer }
        categoryId: { type: integer }
        freelancerId: { type: integer }
        imageUrl: { type: ["string", "null"] }
        tags: { type: ["string", "null"] }
        status: { type: string }
        orderCount: { type: integer }
        rating: { type: number }
        reviewCount: { type: integer }
        isFeatured: { type: boolean }
        createdAt: { type: string }
        category:
          $ref: "#/components/schemas/Category"
        freelancer:
          $ref: "#/components/schemas/FreelancerProfile"
...
    Message:
      type: object
      required: [id, orderId, senderId, content, createdAt]
      properties:
        id: { type: integer }
        orderId: { type: integer }
        senderId: { type: integer }
        content: { type: string }
        attachmentUrl: { type: ["string", "null"] }
        createdAt: { type: string }
        sender:
          $ref: "#/components/schemas/User"
    MessageInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
        attachmentUrl: { type: string }

    # NOTIFICATION
    Notification:
      type: object
      required: [id, userId, type, title, body, isRead, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        type: { type: string }
        title: { type: string }
        body: { type: string }
        link: { type: ["string", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string }

    # ADMIN
...