# lib/api-zod/src/generated/types/categoryUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-SERVICES marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface CategoryUpdate {
  name?: string;
  icon?: string;
  description?: string;
}