# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E-SERVICES marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { FreelancerProfile } from './freelancerProfile';

export interface Service {
  id: number;
  title: string;
  description: string;
  price: number;
  deliveryDays: number;
  categoryId: number;
  freelancerId: number;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  tags?: string | null;
  status: string;
  orderCount: number;
  rating: number;
  reviewCount: number;
  isFeatured?: boolean;
  createdAt: string;
  category?: Category;
  freelancer?: FreelancerProfile;
}